+++ dlrover/python/unified/common/config.py
@@ -10,13 +10,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Dict, Union
+from typing import Dict
 
 from dlrover.python.common.serialize import PickleSerializable
 from dlrover.python.unified.common.constant import DLMasterConstant
 from dlrover.python.unified.common.enums import (
     MasterStateBackendType,
-    RLRoleType,
     SchedulingStrategyType,
 )
 
@@ -31,7 +30,7 @@ class JobConfig(PickleSerializable):
         job_max_restart: int,
         master_max_restart: int,
         trainer_max_restart: int,
-        workload_max_restart: Dict[RLRoleType, int],
+        workload_max_restart: Dict[str, int],
     ):
         """
         Configuration(non-business part) of the job.
@@ -49,7 +48,7 @@ class JobConfig(PickleSerializable):
                 number of master restarts. Default is 10.
             trainer_max_restart (int, optional): The maximum limit on the
                 number of trainer restarts. Default is 10.
-            workload_max_restart (Dict[RLRoleType, int], optional): The
+            workload_max_restart (Dict[str, int], optional): The
                 maximum limit on the number of workload actor restarts.
                 Default is 30.
         """
@@ -109,9 +108,8 @@ class JobConfig(PickleSerializable):
     def workload_max_restart(self):
         return self._workload_max_restart
 
-    def get_workload_max_restart(self, role: Union[str, RLRoleType]):
-        if isinstance(role, str):
-            role = RLRoleType[role.upper()]
+    def get_workload_max_restart(self, role: str):
+        role = role.upper()
 
         if role in self._workload_max_restart:
             return max(self._workload_max_restart[role], self.job_max_restart)
